Create create-macie-job-table-single-job.sql
This option covers creating an Athena table that comprises the output for a single Macie job. This approach is a great starting point to get specific data about a particular job. If you have a scheduled Macie job this table will enable you to query any new sensitive data that is discovered through each scheduled run of a job.