Create macie-object-sensitive-data-location-details-MDI.sql
Provides the location of sensitive data in an object, discovered by Macie MANAGED identifiers. This query is helpful in determining the sensitive data location for multiple objects with findings in a Macie job. Query is written to cover the various locations that Macie can report on depending on the type of file where the sensitive data was discovered.